+++ testprojects/src/scala/org/pantsbuild/testproject/banned_deps_test/Spec.scala
@@ -1,5 +1,3 @@
 package org.pantsbuild.testproject.banned_deps_test
 
-class Spec {
-
-}
+case class Spec()
